+++ tests/ClientTest.php
@@ -9,11 +9,11 @@ declare(strict_types=1);
 
 namespace WebSocket;
 
+use Phrity\Net\Uri;
 use PHPUnit\Framework\TestCase;
 
 class ClientTest extends TestCase
 {
-
     public function setUp(): void
     {
         error_reporting(-1);
@@ -238,19 +238,34 @@ class ClientTest extends TestCase
     public function testBadScheme(): void
     {
         MockSocket::initialize('client.connect', $this);
-        $client = new Client('bad://localhost:8000/my/mock/path');
         $this->expectException('WebSocket\BadUriException');
-        $this->expectExceptionMessage('Url should have scheme ws or wss');
-        $client->send('Connect');
+        $this->expectExceptionMessage("Invalid URI scheme, must be 'ws' or 'wss'.");
+        $client = new Client('bad://localhost:8000/my/mock/path');
     }
 
-    public function testBadUrl(): void
+    public function testBadUri(): void
     {
         MockSocket::initialize('client.connect', $this);
-        $client = new Client('this is not an url');
         $this->expectException('WebSocket\BadUriException');
-        $this->expectExceptionMessage('Invalid url \'this is not an url\' provided.');
+        $this->expectExceptionMessage("Invalid URI '--:this is not an uri:--' provided.");
+        $client = new Client('--:this is not an uri:--');
+    }
+
+    public function testInvalidUriType(): void
+    {
+        MockSocket::initialize('client.connect', $this);
+        $this->expectException('WebSocket\BadUriException');
+        $this->expectExceptionMessage("Provided URI must be a UriInterface or string.");
+        $client = new Client([]);
+    }
+
+    public function testUriInterface(): void
+    {
+        MockSocket::initialize('client.connect', $this);
+        $uri = new Uri('ws://localhost:8000/my/mock/path');
+        $client = new Client($uri);
         $client->send('Connect');
+        $this->assertTrue(MockSocket::isEmpty());
     }
 
     public function testBadStreamContext(): void
@@ -298,7 +313,7 @@ class ClientTest extends TestCase
         $client = new Client('ws://localhost:8000/my/mock/path');
         $this->expectException('WebSocket\ConnectionException');
         $this->expectExceptionCode(0);
-        $this->expectExceptionMessage('Could not read from stream');
+        $this->expectExceptionMessage('Client handshake error');
         $client->send('Connect');
     }
 
@@ -308,7 +323,7 @@ class ClientTest extends TestCase
         $client = new Client('ws://localhost:8000/my/mock/path');
         $this->expectException('WebSocket\ConnectionException');
         $this->expectExceptionCode(0);
-        $this->expectExceptionMessage('Connection to \'ws://localhost/my/mock/path\' failed');
+        $this->expectExceptionMessage('Connection to \'ws://localhost:8000/my/mock/path\' failed');
         $client->send('Connect');
     }
 
@@ -380,6 +395,16 @@ class ClientTest extends TestCase
         $this->expectExceptionCode(1025);
         $this->expectExceptionMessage('Broken frame, read 0 of stated 2 bytes.');
         $client->receive();
+    }
+
+    public function testHandshakeError(): void
+    {
+        MockSocket::initialize('client.connect-handshake-error', $this);
+        $client = new Client('ws://localhost:8000/my/mock/path');
+        $this->expectException('WebSocket\ConnectionException');
+        $this->expectExceptionCode(1024);
+        $this->expectExceptionMessage('Client handshake error');
+        $client->send('Connect');
     }
 
     public function testReadTimeout(): void
@@ -464,7 +489,7 @@ class ClientTest extends TestCase
         MockSocket::initialize('client.connect', $this);
         $client = new Client('ws://localhost:8000/my/mock/path');
         $this->assertNull($client->getName());
-        $this->assertNull($client->getPier());
+        $this->assertNull($client->getRemoteName());
         $this->assertEquals('WebSocket\Client(closed)', "{$client}");
         $client->text('Connect');
         MockSocket::initialize('send-convenicance', $this);
@@ -472,7 +497,7 @@ class ClientTest extends TestCase
         $client->ping();
         $client->pong();
         $this->assertEquals('127.0.0.1:12345', $client->getName());
-        $this->assertEquals('127.0.0.1:8000', $client->getPier());
+        $this->assertEquals('127.0.0.1:8000', $client->getRemoteName());
         $this->assertEquals('WebSocket\Client(127.0.0.1:12345)', "{$client}");
     }
 
@@ -484,7 +509,7 @@ class ClientTest extends TestCase
         $client->close();
         $this->assertFalse($client->isConnected());
         $this->assertNull($client->getName());
-        $this->assertNull($client->getPeer());
+        $this->assertNull($client->getRemoteName());
         $this->assertNull($client->getCloseStatus());
     }
 }